Add files via upload
Bins.py is the pre-processing step to calculate properties of each point, which bin it goes into, its distance and angle from the origin.

Line_of_sight.py does the actual line of sight integration. LoS is a function that has the argument 'func' where func is the integrand at each point along the line. Some 'funcs' are defined below, for example RM and Stokes parameters.